:pencil2: fix: os 사용자수 로직 변경 및 버튼 통계 출력문 삭제

+++ insite-read-service/src/main/java/com/thirty/insitereadservice/activeusers/dto/OsActiveUserDto.java
@@ -25,7 +25,7 @@ public class OsActiveUserDto implements Comparable<OsActiveUserDto>{
 
     public OsActiveUserDto addId(int id,int size){
         this.id = id;
-        this.ratio= (double)count/(double)size;
+        this.ratio= count/(double)size;
         return this;
     }
